[Refactor]: 채팅창 기능별로 컴포넌트 분리 - 아직 완벽히 수정한 것이 아니라 컴파일이 되지 않습니다 - channel에서 EnterRoom 버튼을 누르면 chat page를 띄우도록 구현하고 싶습니다 - SendMessage를 할 때 room 이름을 property로 받아야합니다
